# read instead of browse (3 to 5x faster here)