--- a/addons/sale/sale.py
+++ b/addons/sale/sale.py
@@ -175,8 +175,8 @@ class sale_order(osv.osv):
         return [('id', 'in', [x[0] for x in res])]
 
     def _get_order(self, cr, uid, ids, context=None):
-        lines = self.pool.get('sale.order.line')
-        return list(set(line['order_id'] for line in lines.read(
+        line_obj = self.pool.get('sale.order.line')
+        return list(set(line['order_id'] for line in line_obj.read(
             cr, uid, ids, ['order_id'], load='_classic_write', context=context)))
 
     def _get_default_shop(self, cr, uid, context=None):