OCA#216 for v8