@gdgellatly 2015-05-09
But what if there is more than one id?

I'm not keen on a function signature that indicates it takes multiple ids,
only returning the first result even if today it only ever receives a list
of one.

If performance really mattered then this is the kind of place that the
tradition has been to use direct SQL statements.

On Sun, May 10, 2015 at 12:03 AM, Christophe Combelles <
notifications@github.com> wrote:

> OCA#216 <https://github.com/OCA/OCB/pull/216> for v8
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/OCA/OCB/pull/217
> Commit Summary
>
>    - read instead of browse (3 to 5x faster here)
>
> File Changes
>
>    - *M* addons/sale/sale.py
>    <https://github.com/OCA/OCB/pull/217/files#diff-0> (6)
>
> Patch Links:
>
>    - https://github.com/OCA/OCB/pull/217.patch
>    - https://github.com/OCA/OCB/pull/217.diff
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/OCA/OCB/pull/217>.
>