@ccomb 2015-05-10
Thanks for your comment, I also hate ```[0]``` in openerp code. It's most of the time a source of bugs. However here there is no chance the function receive many ids. I can improve a little bit anyway by merging the different possible results.