@pedrobaeza 2015-05-10
In a Fabien's statement, he said that browse it at least the same fast as read. Why is this difference?