@gdgellatly 2015-05-10
because .id must access two table and two sets of access rules.  Then there
is the setup of the for loop (which can be overcome) To me it is an
optimisation of browse issue.  If .id is requested, don't read another
table, just take it from the existing one.

On Sun, May 10, 2015 at 11:22 PM, Pedro M. Baeza <notifications@github.com>
wrote:

> In a Fabien's statement, he said that browse it at least the same fast as
> read. Why is this difference?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/OCA/OCB/pull/217#issuecomment-100625562>.
>