@ccomb 2015-05-10
I'm sorry to say that I don't trust Fabien's statements: I only trust what I can measure.
It's easy to reproduce: add a timing code around the function, then create a sale order. Then modify a sale order line (change the qty).

On v7 without the patch, the function is run twice:
- 1.98ms
- 1.10ms

With the patch:
- 0.44ms
- 0.45ms

It seems small, but when you run it thousands of times in unittest or in cron operations, it's a real benefit. For a customer I'm also adding caching to this function so that it is only run once. But it needs a different implementation of ormcache (the one found in https://bitbucket.org/anybox/anybox_perf/)