@ccomb 2015-05-10
On v8 the difference is:

without the patch:
- 4.18ms
- 0.11ms

With the patch:
- 1.16ms
- 1.18ms

This is still 2x faster with the patch, however it shows there is a better caching in v8.