@bwrsandman 2015-05-10
You might consider using an OrderedDict since reads are supposed to respect an order. `list(set())` is good for removing duplicates, but the order becomes non-deterministic for `len(ids) > 1`