@ccomb 2015-05-10
@bwrsandman this is not a read, this is a trigger for a function field, which is supposed to return a list of ids to update. The original implementation is based on a dict keys so this not ordered either.