@bwrsandman 2015-05-10
Got it.

I approve the use of read instead of browse for cases which need optimization.
:+1: