@gdgellatly 2015-05-11
Just a question. I am still not sure about this single I'd gurentee.  I
thought if defer computation was set. Like for imports or when field
definitions or defaults change it could receive up to 40 in v7 and loads
more in v8.  But I never really tested it, just something I recall from a
day bug session in an upgrade.
On 11/05/2015 6:11 PM, "Alexandre Fayolle" <notifications@github.com> wrote:

> [image: :+1:]
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/OCA/OCB/pull/217#issuecomment-100782568>.
>