@ccomb 2015-05-11
You're right, I've found the place were the function is called [for v7](https://github.com/OCA/OCB/blob/7.0/openerp/osv/orm.py#L4667) and [for v8](https://github.com/OCA/OCB/blob/8.0/openerp/models.py#L4314).
It clearly shows that the ids directly comes from the ids of the ``write()`` function where the ``_store_get_values`` is called. It's difficult to obtain multiple ids through the web interface because it would require to change many records at once, but it is be possible through the API.
(So my first implementation was indeed broken.)