@pedrobaeza 2015-05-11
But as far as I have seen, you have fixed it, right?