@ccomb 2015-05-11
Yes I did, it now loops on the given order line ids and returns the list of corresponding (and non duplicated)  order_id using a generator expression inside a set.