@pedrobaeza 2015-05-11
I proceed with the merge then.

Thanks.